% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aistats2020smi-package.R
\docType{package}
\name{aistats2020smi-package}
\alias{aistats2020smi-package}
\alias{aistats2020smi}
\title{The 'aistats2020smi' package.}
\description{
Code for the AISTATS 2020 article: Semi-Modular Inference
}
\references{
Chris U. Carmona, Geoff K. Nicholls (2020). Semi-Modular Inference: enhanced learning in multi-modular models by tempering the influence of components
}
