% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmvnorm_arma}
\alias{dmvnorm_arma}
\title{Multivariate Normal Density Function}
\usage{
dmvnorm_arma(x, mean, sigma, logd = FALSE)
}
\arguments{
\item{x}{matrix of observations}

\item{mean}{mean vector}

\item{sigma}{positive definite covariance matrix}

\item{logd}{logical; whether log-density should be returned (default to FALSE)}
}
\description{
This function returns the log-density for a multivariate Gaussian distribution. Originally from: https://gallery.rcpp.org/articles/dmvnorm_arma/
}
