% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_article.R
\name{generate_article}
\alias{generate_article}
\title{\preformatted{Generate Article Carmona & Nicholls (2020)
}}
\usage{
generate_article(
  out_dir = ".",
  compute_mcmc = FALSE,
  mcmc_dir = NULL,
  compile_path = NULL,
  pkg_dir = system.file(package = "aistats2020smi")
)
}
\arguments{
\item{out_dir}{Output directory where the pdf and auxiliary latex files will be stored}

\item{compute_mcmc}{Indicates if the MCMC must be recomputed. If TRUE, the computation time can be in the order of several hours.}

\item{mcmc_dir}{Directory containing the required files with MCMC results.}

\item{compile_path}{Directory where the article is compiled and all auxiliary files are saved into. If NULL, a temporary directory is used.}

\item{pkg_dir}{Internal directory where the package is installed. Used to retrieve the master Rnw file.}
}
\value{
The function will generate a pdf corresponding to the statistical supplement of Styring et al. (2020)
}
\description{
\preformatted{This function reproduce the article ``Semi-Modular Inference: enhanced learning in multi-modular models by tempering the influence of components''.
The function copies an internal .Rnw file to the specified output folder, and then compiles it using the knitr package.
}
}
