% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_agricurb_data.R
\name{get_agricurb_data}
\alias{get_agricurb_data}
\title{\preformatted{Function to obtain the real agricultural data in a convenient format
}}
\usage{
get_agricurb_data(
  arc_datasets = "NMeso",
  vars_log = NULL,
  vars_sqrt = NULL,
  vars_scale_int_0_1 = NULL,
  vars_scale_mean_0_var_1 = NULL
)
}
\arguments{
\item{arc_datasets}{indicated the datasets used for sampling the archaeological covariates.}

\item{vars_log}{indicated the variables to be rescale to have mean \code{0} and variance \code{1}.}

\item{vars_sqrt}{indicated the variables that are transformed with square root.}

\item{vars_scale_int_0_1}{indicated the variables to be rescale to the interval \code{[0,1]}.}

\item{vars_scale_mean_0_var_1}{indicated the variables to be rescale to have mean \code{0} and variance \code{1}.}
}
\value{
Returns a data frame similar to AgricurbAll ready to analyse with the current methods, such as \code{mcmc_PO1_HM1_HM1}.
}
\description{
\preformatted{The function \code{get_agricurb_data} pre-process the agricultural data into a convenient format ready to analisis.

The formatting process is as follows:
}

\enumerate{
\item Keep modern data and as many archaeological datasets as specified in \code{arc_datasets}.
\item Eliminate rows with unknown \code{Size} in the archaeological datasets.
\item Eliminate rows with \code{Category} different to \code{wheat} and \code{barley}.
\item Modify \code{Rainfall} in the archaeological data sets. Creates \code{Rainfall_min=Rainfall+minRainfall} and \code{Rainfall_max=Rainfall+maxRainfall}, and after that \code{Rainfall=NA}.
\item Transform the variables listed in \code{vars_log} to a logaritmic scale.
\item Rescale the variables listed in \code{vars_scale_int_0_1} to the interval \code{[0,1]}.
\item Rescale the variables listed in \code{vars_scale_mean_0_var_1} to have empirical mean 0 and empirical variance 1.
}
}
\examples{

require(ggplot2)

Agricurb_data <- get_agricurb_data( arc_datasets = "NMeso",
                                         vars_log = c("Rainfall"),
                                         vars_scale_int_0_1 = c("Size", "Rainfall"),
                                         vars_scale_mean_0_var_1 = NULL )
Agricurb_data$ManureLevel <- factor(Agricurb_data$ManureLevel,levels=c("low","medium","high"))

# archaeological data identifier
id_arch <- !is.element(Agricurb_data$dataset,"modern")

#####
# Checking data characteristics
#####

# HM model #
# Relation between Rainfall and normd15N
# (assuming missing data)
p <- ggplot( aes(x=Rainfall,y=normd15N), data=Agricurb_data)
p <- p + geom_point(aes(colour=ManureLevel), pch=20, size=3, alpha=0.7)
p

}
