% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpv_loglik.R
\name{hpv_loglik}
\alias{hpv_loglik}
\title{Joint log-likelihood for the HPV model}
\usage{
hpv_loglik(
  data,
  theta,
  phi,
  eta_modules = stats::setNames(c(1, 1), c("pois", "binom"))
)
}
\arguments{
\item{data}{data frame with n_obs rows and four columns: "nhpv", "Npart", "ncases", "Npop"}

\item{theta}{matrix of dimension c(2, 1). Two parameters defining the linear relation between HPV prevalence and cancer incidence in the HPV model.}

\item{phi}{matrix of dimension c(n_obs, 1). The HPV prevalence parameter of each population. values in (0,1)}

\item{eta_modules}{Named tuple specifying the degree of influence for the "pois" and "binom" modules in the HPV model.}
}
\description{
Computes the joint log-likelihood for the HPV model
}
