% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_misc.R
\name{set_ggtheme}
\alias{set_ggtheme}
\title{Set default ggplot2 theme}
\usage{
set_ggtheme(resize = FALSE)
}
\arguments{
\item{resize}{Resize some of the text elements}
}
\description{
Set default ggplot2 theme

Configures ggplot2 theme to be used in our plots
}
