% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smi_sec_5_1.R
\name{smi_sec_5_1_expectation}
\alias{smi_sec_5_1_expectation}
\title{SMI on synthetic data, on expectation}
\usage{
smi_sec_5_1_expectation(
  out_dir,
  n_iter = 1000,
  n_new = 1000,
  phi = 0,
  theta = 1,
  n = 25,
  m = 50,
  sigma_z = 2,
  sigma_y = 1,
  sigma_phi = Inf,
  sigma_theta = 0.5,
  eta_all = seq(0, 1, 0.025)
)
}
\arguments{
\item{out_dir}{Directory where outputs are stored}

\item{n_iter}{number of iterations}

\item{n_new}{number of iterations}

\item{phi}{Generative (true) value assumed for phi}

\item{theta}{Generative (true) value assumed for theta}

\item{n}{Sample size for Z}

\item{m}{Sample size for Y}

\item{sigma_z}{Likelihood variance for Z}

\item{sigma_y}{Likelihood variance for Y}

\item{sigma_phi}{Prior std dev phi}

\item{sigma_theta}{Prior std dev eta}

\item{eta_all}{Degree of influence of the unbiased module into phi}
}
\description{
Illustrate convenience of SMI using Monte Carlo integration over the data generation
}
