% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smi_sec_5_1.R
\name{smi_sec_5_1_single_dataset}
\alias{smi_sec_5_1_single_dataset}
\title{SMI on synthetic data, single dataset}
\usage{
smi_sec_5_1_single_dataset(
  out_dir,
  phi = 0,
  theta = 1,
  n = 25,
  m = 50,
  sigma_z = 2,
  sigma_y = 1,
  sigma_phi = Inf,
  sigma_theta = 0.5,
  eta_all = seq(0, 1, 0.025)
)
}
\arguments{
\item{out_dir}{Directory where outputs are stored}

\item{phi}{Generative (true) value assumed for phi}

\item{theta}{Generative (true) value assumed for theta}

\item{n}{Sample size for Z}

\item{m}{Sample size for Y}

\item{sigma_z}{Likelihood variance for Z}

\item{sigma_y}{Likelihood variance for Y}

\item{sigma_phi}{Prior std dev phi}

\item{sigma_theta}{Prior std dev theta}

\item{eta_all}{Degree of influence of the unbiased module into phi}
}
\description{
Illustrate convenience of SMI on a randomly generated dataset
}
