% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smi_sec_5_3.R
\name{smi_sec_5_3_compute_mcmc}
\alias{smi_sec_5_3_compute_mcmc}
\title{Compute MCMC in agricultural model}
\usage{
smi_sec_5_3_compute_mcmc(
  out_dir,
  eta_all = seq(0.1, 1, by = 0.1),
  force_compute_mcmc = FALSE,
  n_iter = 1e+05,
  n_iter_sub = 100,
  n_warmup = 10000,
  n_thin = 3,
  n_epoch_adapt = 5,
  n_iter_adapt = 1000,
  seed = 123
)
}
\arguments{
\item{out_dir}{Directory where outputs are stored}

\item{eta_all}{Levels of influence of the PO module}

\item{force_compute_mcmc}{To avoid computation time, completed MCMC is not recomputed, unless force_compute_mcmc is True}

\item{n_iter}{Number of iterations in the main chain}

\item{n_iter_sub}{Number of iterations in the sub-chain for the two-stage MCMC}

\item{n_warmup}{Number of iterations used for warm-up}

\item{n_thin}{Number of iterations used for thinning}

\item{n_epoch_adapt}{Number of epochs for adaptation of proposal distributions}

\item{n_iter_adapt}{Number of iterations in each adaptation epoch}

\item{seed}{Set random seed before MCMC starts for reproducibility}
}
\description{
Compute MCMC used in section 5.3
}
