% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smi_sec_5_4.R
\name{smi_sec_5_4_compute_mcmc}
\alias{smi_sec_5_4_compute_mcmc}
\title{Compute MCMC in HPV model}
\usage{
smi_sec_5_4_compute_mcmc(
  out_dir,
  eta_all = seq(0.1, 1, by = 0.1),
  force_compute_mcmc = FALSE,
  n_iter = 5000,
  n_iter_sub = 500,
  n_warmup = 1000,
  n_cores = 8,
  seed = 123,
  pkg_dir = system.file(package = "aistats2020smi")
)
}
\arguments{
\item{out_dir}{Directory where outputs are stored}

\item{eta_all}{Levels of influence of the Poisson module}

\item{force_compute_mcmc}{To avoid computation time, completed MCMC is not recomputed, unless force_compute_mcmc is True}

\item{n_iter}{Number of iterations in the main chain}

\item{n_iter_sub}{Number of iterations in the sub-chain for the two-stage MCMC}

\item{n_warmup}{Number of iterations used for warm-up}

\item{n_cores}{Number of cores used by stan}

\item{seed}{Set random seed before MCMC starts for reproducibility}

\item{pkg_dir}{Internal directory where the package is installed. Used to retrieve stan scripts for the HPV model.}
}
\description{
Compute MCMC used in section 5.3
}
