% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smi_sec_5_4.R
\name{smi_sec_5_4_compute_summary}
\alias{smi_sec_5_4_compute_summary}
\title{Compute summary of MCMC results in HPV model}
\usage{
smi_sec_5_4_compute_summary(mcmc_dir, out_dir, eta_all)
}
\arguments{
\item{mcmc_dir}{Directory where the MCMC results are read from}

\item{out_dir}{Directory where outputs are saved}

\item{eta_all}{Levels of influence of the Poisson module}
}
\description{
Compute summary of MCMC used in section 5.4
}
