% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smi_sec_5_4.R
\name{smi_sec_5_4_eta_selection}
\alias{smi_sec_5_4_eta_selection}
\title{Compute eta selection in HPV model}
\usage{
smi_sec_5_4_eta_selection(mcmc_dir, out_dir, eta_all = seq(0.1, 1, by = 0.1))
}
\arguments{
\item{mcmc_dir}{Directory where the MCMC results are read from}

\item{out_dir}{Directory where outputs are saved}

\item{eta_all}{Levels of influence of the poisson module}
}
\description{
Compute eta selection in HPV model for section 5.4
}
