% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smi_sec_5_4.R
\name{smi_sec_5_4_plot_eta_selection}
\alias{smi_sec_5_4_plot_eta_selection}
\title{Plotting model evaluation results in HPV model}
\usage{
smi_sec_5_4_plot_eta_selection(mcmc_dir, out_dir)
}
\arguments{
\item{mcmc_dir}{Directory where the MCMC results are read from}

\item{out_dir}{Directory where outputs are stored}
}
\description{
Plot model evaluation results in HPV model for section 5.4
}
