% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smi_sec_5_4.R
\name{smi_sec_5_4_plot_mcmc}
\alias{smi_sec_5_4_plot_mcmc}
\title{Plot MCMC results in HPV model}
\usage{
smi_sec_5_4_plot_mcmc(mcmc_dir, out_dir)
}
\arguments{
\item{mcmc_dir}{Directory with MCMC summaries produced by smi_sec_5_4_compute_summary}

\item{out_dir}{Directory where outputs are saved}
}
\description{
Plot MCMC results from section 5.4
}
